requirements.txt
pip
unstructured
networkx
pyroscope-io
python-telegram-bot
sqlalchemy
python-dotenv
psycopg2